+++ modmesh/__init__.py
@@ -41,8 +41,6 @@ from . import spacetime  # noqa: F401
 from . import onedim  # noqa: F401
 from . import system  # noqa: F401
 from . import toggle  # noqa: F401
-from . import timeseries_dataframe  # noqa: F401, F403
-from .timeseries_dataframe import *  # noqa: F401, F403
 
 
 clinfo = core.ProcessInfo.instance.command_line
